requirements.txt
asgiref==3.7.2
Django==4.2.3
django-environ==0.10.0
djangorestframework==3.14.0
gunicorn==21.2.0
mysqlclient==2.1.0
packaging==23.1
pytz==2023.3
sqlparse==0.4.4
